separated flight filters into a component
